[jcommon] Use long for TimeInterval.length in jackson serialization/deserialization
Summary:
SAT, TimeInterval.length may exceed max int value.

I'm listing the reviewers of D562444 for this diff -- please let me know if more people should be included.

Test Plan: added unit test, existing tests pass

Reviewers: groys, timw, srash, dain, ajaiswal

Reviewed By: groys

CC: security-diffs@lists

Differential Revision: https://phabricator.fb.com/D636817

+++ util/src/main/java/com/facebook/util/TimeInterval.java
@@ -99,7 +99,7 @@ public class TimeInterval {
   @JsonCreator
   private static TimeInterval fromJson(
     @JsonProperty("type") TimeIntervalType type,
-    @JsonProperty("length") int length
+    @JsonProperty("length") long length
   ) {
     if (type == null) {
       if (length == 0) {
